+++ Rubrik/Private/Get-RubrikAPIData.ps1
@@ -3392,8 +3392,8 @@ function Get-RubrikAPIData {
                 URI         = '/api/internal/config/usersettable_volumeGroup'
                 Method      = 'Patch'
                 Body        = @{
-                      migrateFastVirtualDiskBuild = 'migrateFastVirtualDiskBuild'
-                      maxFullMigrationStoragePercentage = 'maxFullMigrationStoragePercentage'
+                    migrateFastVirtualDiskBuild = 'migrateFastVirtualDiskBuild'
+                    maxFullMigrationStoragePercentage = 'maxFullMigrationStoragePercentage'
                 }
                 Query       = ''
                 Result      = ''
@@ -3531,6 +3531,23 @@ function Get-RubrikAPIData {
                 Result      = ''
                 Filter      = ''
                 Success     = '202'
+            }
+        }
+        'Test-RubrikSnapshotVerification' = @{
+            '5.3' = @{
+                Description = 'Verifies the validity of a snapshot'
+                URI         = '/api/v1/backup/verify'
+                Method      = 'Post'
+                Body        = @{
+                    "objectId" = ''
+                    "snapshotIdsOpt" = [System.Collections.ArrayList]@()
+                    "locationIdOpt" = "locationIdOpt"
+                    "shouldVerifyAfterOpt" = 'shouldVerifyAfterOpt'
+                }
+                Query       = ''
+                Result      = ''
+                Filter      = ''
+                Success     = '200'
             }
         }
         'Update-RubrikVCenter'         = @{